Fixed the session chooser to start selection
The session choosing options will now highlight the default session
as defined in lightdm.conf (lightdm.default_session is the key, it
cannot be accessed with lightdm.default_session.key).

Additionally, the selected option will now be considered when
calling lightdm.login instead of passing the default session; if
any of the radio buttons is selected, the corresponding key will be
used, otherwise, lightdm.default_session is passed as before.